refactor(codegen): refactor utils
